Added FileDealer and changed blockSize from long to int

+++ src/main/java/ciphers/AesCipher.java
@@ -110,7 +110,7 @@ public class AesCipher  implements ICipher
         }
     }
 
-    public long getBlockSize()
+    public int getBlockSize()
     {
         return 16;
     }
